Cleanup Initial
File reconstructure for client
