+++ src/client/headers/msg.h
@@ -1,0 +1,8 @@
+#ifndef MSG_H
+#define MSG_H
+
+void set_name(int s);
+void send_msg(int s, char *m);
+void listen_messages(int s);
+
+#endif // MSG_H
